# swagger.yaml
openapi: 3.0.3
info:
  title: Playlists API
  version: 1.0.0
tags:
  - name: playlists
    description: Playlists management
  - name: tracks
    description: Tracks management
  - name: user
    description: User management
  - name: search
    description: Search system
paths:
  /playlists:
    get: # get all public playlists
      tags:
        - playlists
      summary: Get all playlists
      operationId: getPlaylists
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Playlist'
    post: # create playlist
      tags:
        - playlists
      summary: Create playlist
      operationId: createPlaylist
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                is_public:
                  type: boolean
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playlist'
      security:
        - bearerAuth: []
  /playlists/{playlistId}:
    parameters:
      - name: playlistId
        in: path
        description: The id of the playlist
        required: true
        schema:
          type: integer
    get: # get playlist
      tags:
        - playlists
      summary: Get a playlist
      operationId: getPlaylist
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playlist'
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - bearerAuth: []
    delete: # remove playlist
      tags:
        - playlists
      summary: Delete a playlist
      operationId: deletePlaylist
      responses:
        '204':
          description: No Content
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - bearerAuth: []
    put: # create/overwrite playlist
      tags:
        - playlists
      summary: Update a playlist
      operationId: updatePlaylist
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                is_public:
                  type: boolean
              
        required: true
      responses:
        '201':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playlist'
        '400':
          description: Bad Request
        '403':
          description: Forbidden
      security:
        - bearerAuth: []
  /playlists/{playlistId}/tracks:
    parameters:
      - name: playlistId
        in: path
        description: Playlist ID
        required: true
        schema:
          type: integer
    get:
      tags:
        - playlists
      summary: Get all tracks from a playlist
      operationId: getAllPlaylistTracks
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrackInPlaylist'
        '400':
          description: Bad request
        '404':
          description: Playlist not found
    post: # add track to playlist
      tags:
        - playlists
      summary: Add a track to a playlist
      operationId: addTrackToPlaylist
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                youtube_id:
                  type: string
                  example: "_E3FGh4b04s"
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playlist'
        '400':
          description: Bad Request
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - bearerAuth: []
  /playlists/{playlistId}/tracks/{youtube_id}:
    parameters:
      - name: playlistId
        in: path
        description: Playlist ID
        required: true
        schema:
          type: integer
      - name: youtube_id
        in: path
        description: Track ID
        required: true
        schema:
          type: string
    delete: # remove track from playlist (check is_public and owner)
      tags:
        - playlists
      summary: Delete a track from a playlist
      operationId: deletePlaylistTrack
      responses:
        '204':
          description: No Content
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - bearerAuth: []

  /tracks/{trackId}:
    parameters:
      - name: trackId
        in: path
        description: Track ID
        required: true
        schema:
          type: string
    get: # get track
      tags:
        - tracks
      summary: Get a track
      operationId: getTrack
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '404':
          description: Not Found

  /user:
    post: # create user
      tags:
        - user
      summary: Create a user
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: Sanjo
                password:
                  type: string
                  example: sd355ohf2348d
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad Request
  /user/login:
    post: # login user
      tags:
        - user
      summary: Login a user
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: Sanjo
                password:
                  type: string
                  example: sd355ohf2348d
      responses:
        '200':
          description: OK 
          headers:
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Invalid username/password provided
  /user/logout:
    post: # logout user
      tags:
        - user
      summary: Logout a user
      operationId: logoutUser
      responses:
        '200':
          description: OK
        '400':
          description: Invalid token provided
      security:
        - bearerAuth: []
  /user/{userId}:
    parameters:
      - name: userId
        in: path
        description: The id of the user
        required: true
        schema:
          type: integer
    get: # get user
      tags:
        - user
      summary: Get a user
      operationId: getUser
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Not Found
    put: # update user data
      tags:
        - user
      summary: Edit user data
      operationId: editUser
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: Sanjo
                password:
                  type: string
                  example: sd355ohf2348d
                newPassword:
                  type: string
                  example: sd355ohf2348d
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad Request
        '403':
          description: Forbidden
        '404': 
          description: Not Found
      security:
        - bearerAuth: []
    delete: # remove user
      tags:
        - user
      summary: Delete a user
      operationId: deleteUser
      parameters:
        - name: password
          in: query
          description: Hash of password
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No Content
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - bearerAuth: []
  /search/tracks:
    get: # search tracks
      tags:
        - search
      summary: Search tracks
      operationId: searchTracks
      parameters:
        - name: query
          in: query
          description: Search query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Track'
  /search/playlists:
    get: # search playlists
      tags:
        - search
      summary: Search playlists
      operationId: searchPlaylists
      parameters:
        - name: query
          in: query
          description: Search query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Playlist'
  /search/users:
    get: # search users
      tags:
        - search
      summary: Search users
      operationId: searchUsers
      parameters:
        - name: query
          in: query
          description: Search query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    TrackInPlaylist:
      type: object
      properties:
        title:
          type: string
          example: KORDHELL - KILLERS FROM THE NORTHSIDE
        playlist_id:
          type: integer
          example: 1
        youtube_id:
          type: string
          example: DK6IRG4CAbw
    Track:
      type: object
      properties:
        title:
          type: string
          example: KORDHELL - KILLERS FROM THE NORTHSIDE
        youtube_id:
          type: string
          example: DK6IRG4CAbw
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: roflanmen
        playlists:
          type: array
          items: 
            $ref: '#/components/schemas/Playlist'
    Playlist:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: My playlist
        is_public:
          type: boolean
          example: true
        owner_id:
          type: integer
          example: 1
        tracks:
          type: array
          items: 
            $ref: '#/components/schemas/TrackInPlaylist'